+++ Mailer/app/Forms/DailyMinuteTemplateFormFactory.php
@@ -91,7 +91,6 @@ class DailyMinuteTemplateFormFactory
             ->setRequired("Field 'Type' is required.");
 
         $form->addText('from', 'Sender')
-            ->setHtmlAttribute('placeholder', 'e.g. info@domain.com')
             ->setRequired("Field 'Sender' is required.");
 
         $form->addText('subject', 'Subject')