+++ Mailer/app/Forms/DennikeTemplateFormFactory.php
@@ -85,7 +85,6 @@ class DennikeTemplateFormFactory
             ->setRequired("Field 'Type' is required.");
 
         $form->addText('from', 'Sender')
-            ->setHtmlAttribute('placeholder', 'e.g. info@domain.com')
             ->setRequired("Field 'Sender' is required.");
 
         $form->addText('subject', 'Subject')