+++ Mailer/app/Forms/EuobserverTemplateFormFactory.php
@@ -58,7 +58,6 @@ class EuobserverTemplateFormFactory
             ->setRequired("Field 'Type' is required.");
 
         $form->addText('from', 'Sender')
-            ->setHtmlAttribute('placeholder', 'e.g. info@domain.com')
             ->setRequired("Field 'Sender' is required.");
 
         $form->addText('subject', 'Subject')