+++ Mailer/app/Forms/GrafdnaTemplateFormFactory.php
@@ -57,7 +57,6 @@ class GrafdnaTemplateFormFactory
             ->setRequired("Field 'Type' is required.");
 
         $form->addText('from', 'Sender')
-            ->setHtmlAttribute('placeholder', 'e.g. info@domain.com')
             ->setRequired("Field 'Sender' is required.");
 
         $form->addText('subject', 'Subject')