+++ Mailer/app/Forms/MediaBriefingTemplateFormFactory.php
@@ -86,7 +86,6 @@ class MediaBriefingTemplateFormFactory
             ->setRequired("Field 'Type' is required.");
 
         $form->addText('from', 'Sender')
-            ->setHtmlAttribute('placeholder', 'e.g. info@domain.com')
             ->setRequired("Field 'Sender' is required.");
 
         $form->addText('subject', 'Subject')