+++ Mailer/app/Forms/RespektArticleParserTemplateFormFactory.php
@@ -100,7 +100,6 @@ class RespektArticleParserTemplateFormFactory
             ->setRequired("Field 'Type' is required.");
 
         $form->addText('from', 'Sender')
-            ->setHtmlAttribute('placeholder', 'e.g. info@domain.com')
             ->setRequired("Field 'Sender' is required.");
 
         $form->addText('subject', 'Subject')